Refactored sca inventory sca and added some types
