+++ public/components/agents/sca/inventory.tsx
@@ -24,8 +24,6 @@ import {
   EuiButton,
   EuiButtonIcon,
   EuiStat,
-  EuiHealth,
-  EuiDescriptionList,
   EuiButtonEmpty,
   EuiToolTip,
   EuiCallOut,
@@ -36,8 +34,6 @@ import { WzRequest } from '../../../react-services/wz-request';
 import { formatUIDate } from '../../../react-services/time-service';
 import exportCsv from '../../../react-services/wz-csv';
 import { getToasts } from '../../../kibana-services';
-import { WzSearchBar } from '../../../components/wz-search-bar';
-import { RuleText, ComplianceText } from './components';
 import _ from 'lodash';
 import {
   UI_ERROR_SEVERITIES,
@@ -48,24 +44,42 @@ import {
 import { API_NAME_AGENT_STATUS, UI_LOGGER_LEVELS } from '../../../../common/constants';
 import { getErrorOrchestrator } from '../../../react-services/common-services';
 import { VisualizationBasic } from '../../common/charts/visualizations/basic';
-import { TableWzAPI } from '../../common/tables';
-import { getFilterValues } from './lib/api-request';
-export class Inventory extends Component {
+import { InventoryPolicyChecksTable } from './inventory/checks-table';
+
+type InventoryProps = {
+  agent: { [key: string]: any };
+};
+
+type InventoryState = {
+  agent: object;
+  itemIdToExpandedRowMap: object;
+  showMoreInfo: boolean;
+  loading: boolean;
+  filters: object[];
+  pageTableChecks: { pageIndex: number; pageSize?: number };
+  policies: object[];
+  lookingPolicy: { [key: string]: any } | false;
+  loadingPolicy: boolean;
+};
+export class Inventory extends Component<InventoryProps, InventoryState> {
   _isMount = false;
+  columnsPolicies: object[];
+  lookingPolicy: { [key: string]: any } | false = false;
   constructor(props) {
     super(props);
     const { agent } = this.props;
     this.state = {
       agent,
-      items: [],
       itemIdToExpandedRowMap: {},
       showMoreInfo: false,
       loading: false,
       filters: [],
       pageTableChecks: { pageIndex: 0 },
       policies: [],
+      lookingPolicy: false,
+      loadingPolicy: false,
     };
-    this.suggestions = {};
+
     this.columnsPolicies = [
       {
         field: 'name',
@@ -76,7 +90,6 @@ export class Inventory extends Component {
         field: 'description',
         name: 'Description',
         truncateText: true,
-        render: formatUIDate,
         sortable: true,
       },
       {
@@ -111,71 +124,6 @@ export class Inventory extends Component {
           return `${score}%`;
         },
         width: '100px',
-      },
-    ];
-    this.columnsChecks = [
-      {
-        field: 'id',
-        name: 'ID',
-        sortable: true,
-        width: '100px',
-      },
-      {
-        field: 'title',
-        name: 'Title',
-        sortable: true,
-        truncateText: true,
-      },
-      {
-        name: 'Target',
-        truncateText: true,
-        render: (item) => (
-          <div>
-            {item.file ? (
-              <span>
-                <b>File:</b> {item.file}
-              </span>
-            ) : item.directory ? (
-              <span>
-                <b>Directory:</b> {item.directory}
-              </span>
-            ) : item.process ? (
-              <span>
-                <b>Process: </b> {item.process}
-              </span>
-            ) : item.command ? (
-              <span>
-                <b>Command: </b> {item.command}
-              </span>
-            ) : item.registry ? (
-              <span>
-                <b>Registry: </b> {item.registry}
-              </span>
-            ) : (
-              '-'
-            )}
-          </div>
-        ),
-      },
-      {
-        field: 'result',
-        name: 'Result',
-        truncateText: true,
-        sortable: true,
-        width: '150px',
-        render: this.addHealthResultRender,
-      },
-      {
-        align: 'right',
-        width: '40px',
-        isExpander: true,
-        render: (item) => (
-          <EuiButtonIcon
-            onClick={() => this.toggleDetails(item)}
-            aria-label={this.state.itemIdToExpandedRowMap[item.id] ? 'Collapse' : 'Expand'}
-            iconType={this.state.itemIdToExpandedRowMap[item.id] ? 'arrowUp' : 'arrowDown'}
-          />
-        ),
       },
     ];
   }
@@ -232,134 +180,6 @@ export class Inventory extends Component {
     this._isMount = false;
   }
 
-  addHealthResultRender(result) {
-    const color = (result) => {
-      if (result.toLowerCase() === 'passed') {
-        return 'success';
-      } else if (result.toLowerCase() === 'failed') {
-        return 'danger';
-      } else {
-        return 'subdued';
-      }
-    };
-
-    return (
-      <EuiHealth color={color(result)} style={{ textTransform: 'capitalize' }}>
-        {result || 'Not applicable'}
-      </EuiHealth>
-    );
-  }
-
-  /**
-   * Generate and assign the suggestions for the searchbar
-   * @param policy
-   * @param checks
-   * @returns
-   */
-  buildSuggestionSearchBar(policy, checks) {
-    if (this.suggestions[policy]) return;
-    const distinctFields = {};
-    checks.forEach((item) => {
-      Object.keys(item).forEach((field) => {
-        if (typeof item[field] === 'string') {
-          if (!distinctFields[field]) {
-            distinctFields[field] = {};
-          }
-          if (!distinctFields[field][item[field]]) {
-            distinctFields[field][item[field]] = true;
-          }
-        }
-      });
-    });
-
-    /**
-     * Get list of suggestions.
-     * This method validate if the suggestion item exists in the checks array fields
-     * @returns List of suggestions
-     */
-    const getSuggestionsFields = (policy) => {
-      const defaultSuggestions = [
-        {
-          type: 'params',
-          label: 'condition',
-          description: 'Filter by check condition',
-          operators: ['=', '!='],
-          values: (value) => getFilterValues('condition', value, this.props.agent.id, policy),
-        },
-        {
-          type: 'params',
-          label: 'file',
-          description: 'Filter by check file',
-          operators: ['=', '!='],
-          values: (value) => getFilterValues('file', value, this.props.agent.id, policy),
-        },
-        {
-          type: 'params',
-          label: 'title',
-          description: 'Filter by check title',
-          operators: ['=', '!='],
-          values: (value) => getFilterValues('title', value, this.props.agent.id, policy),
-        },
-        {
-          type: 'params',
-          label: 'result',
-          description: 'Filter by check result',
-          operators: ['=', '!='],
-          values: (value) => getFilterValues('result', value, this.props.agent.id, policy),
-        },
-        {
-          type: 'params',
-          label: 'status',
-          description: 'Filter by check status',
-          operators: ['=', '!='],
-          values: (value) => getFilterValues('status', value, this.props.agent.id, policy),
-        },
-        {
-          type: 'params',
-          label: 'rationale',
-          description: 'Filter by check rationale',
-          operators: ['=', '!='],
-          values: (value) => getFilterValues('rationale', value, this.props.agent.id, policy),
-        },
-        {
-          type: 'params',
-          label: 'registry',
-          description: 'Filter by check registry',
-          operators: ['=', '!='],
-          values: (value) => getFilterValues('registry', value, this.props.agent.id, policy),
-        },
-        {
-          type: 'params',
-          label: 'description',
-          description: 'Filter by check description',
-          operators: ['=', '!='],
-          values: (value) => getFilterValues('description', value, this.props.agent.id, policy),
-        },
-        {
-          type: 'params',
-          label: 'remediation',
-          description: 'Filter by check remediation',
-          operators: ['=', '!='],
-          values: (value) => getFilterValues('remediation', value, this.props.agent.id, policy),
-        },
-        {
-          type: 'params',
-          label: 'reason',
-          description: 'Filter by check reason',
-          operators: ['=', '!='],
-          values: (value) => getFilterValues('reason', value, this.props.agent.id, policy),
-        },
-      ];
-
-      const filteresSuggestions = defaultSuggestions.filter((item) =>
-        Object.keys(distinctFields).includes(item.label)
-      );
-      return filteresSuggestions;
-    };
-
-    this.suggestions[policy] = getSuggestionsFields(policy);
-  }
-
   async initialize() {
     try {
       this._isMount && this.setState({ loading: true });
@@ -387,6 +207,10 @@ export class Inventory extends Component {
     }
   }
 
+  /**
+   *
+   * @param policy
+   */
   async loadScaPolicy(policy) {
     this._isMount &&
       this.setState({
@@ -402,22 +226,9 @@ export class Inventory extends Component {
           },
         });
         const [policyData] = policyResponse.data.data.affected_items;
-        // It queries all checks without filters, because the filters are applied in the results
-        // due to the use of EuiInMemoryTable instead EuiTable components and do arequest with each change of filters.
-        const checksResponse = await WzRequest.apiReq(
-          'GET',
-          `/sca/${this.props.agent.id}/checks/${policy.policy_id}`,
-          {}
-        );
-        const checks = (
-          (((checksResponse || {}).data || {}).data || {}).affected_items || []
-        ).map((item) => ({ ...item, result: item.result || 'not applicable' }));
-        this.buildSuggestionSearchBar(policyData.policy_id, checks);
-        this._isMount &&
-          this.setState({ lookingPolicy: policyData, loadingPolicy: false, items: checks });
+        this._isMount && this.setState({ lookingPolicy: policyData, loadingPolicy: false });
       } catch (error) {
         this.setState({ lookingPolicy: policy, loadingPolicy: false });
-
         const options: UIErrorLog = {
           context: `${Inventory.name}.loadScaPolicy`,
           level: UI_LOGGER_LEVELS.ERROR as UILogLevel,
@@ -435,58 +246,12 @@ export class Inventory extends Component {
     }
   }
 
-  toggleDetails = (item) => {
-    const itemIdToExpandedRowMap = { ...this.state.itemIdToExpandedRowMap };
-
-    if (itemIdToExpandedRowMap[item.id]) {
-      delete itemIdToExpandedRowMap[item.id];
-    } else {
-      let checks = '';
-      checks += (item.rules || []).length > 1 ? 'Checks' : 'Check';
-      checks += item.condition ? ` (Condition: ${item.condition})` : '';
-      const complianceText =
-        item.compliance && item.compliance.length
-          ? item.compliance.map((el) => `${el.key}: ${el.value}`).join('\n')
-          : '';
-      const rulesText = item.rules.length ? item.rules.map((el) => el.rule).join('\n') : '';
-      const listItems = [
-        {
-          title: 'Check not applicable due to:',
-          description: item.reason,
-        },
-        {
-          title: 'Rationale',
-          description: item.rationale || '-',
-        },
-        {
-          title: 'Remediation',
-          description: item.remediation || '-',
-        },
-        {
-          title: 'Description',
-          description: item.description || '-',
-        },
-        {
-          title: (item.directory || '').includes(',') ? 'Paths' : 'Path',
-          description: item.directory,
-        },
-        {
-          title: checks,
-          description: <RuleText rules={item.rules.length ? item.rules : []} />,
-        },
-        {
-          title: 'Compliance',
-          description: <ComplianceText complianceText={complianceText} />,
-        },
-      ];
-      const itemsToShow = listItems.filter((x) => {
-        return x.description;
-      });
-      itemIdToExpandedRowMap[item.id] = <EuiDescriptionList listItems={itemsToShow} />;
-    }
-    this.setState({ itemIdToExpandedRowMap });
-  };
-
+  /**
+   *
+   * @param color
+   * @param title
+   * @param time
+   */
   showToast = (color, title, time) => {
     getToasts().add({
       color: color,
@@ -494,6 +259,10 @@ export class Inventory extends Component {
       toastLifeTimeMs: time,
     });
   };
+
+  /**
+   *
+   */
   async downloadCsv() {
     try {
       this.showToast('success', 'Your download should begin automatically...', 3000);
@@ -532,20 +301,6 @@ export class Inventory extends Component {
         className: 'customRowClass',
         onClick: () => this.loadScaPolicy(item),
       };
-    };
-    const getChecksRowProps = (item, idx) => {
-      return {
-        'data-test-subj': `sca-check-row-${idx}`,
-        className: 'customRowClass',
-        onClick: () => this.toggleDetails(item),
-      };
-    };
-
-    const sorting = {
-      sort: {
-        field: 'id',
-        direction: 'asc',
-      },
     };
     const buttonPopover = (
       <EuiButtonEmpty
@@ -597,7 +352,7 @@ export class Inventory extends Component {
               <div>
                 {this.state.policies.length && (
                   <EuiFlexGroup style={{ marginTop: 0 }}>
-                    {this.state.policies.map((policy, idx) => (
+                    {this.state.policies.map((policy: any, idx) => (
                       <EuiFlexItem key={idx} grow={false}>
                         <EuiCard
                           title
@@ -681,22 +436,6 @@ export class Inventory extends Component {
                         </h2>
                       </EuiTitle>
                     </EuiFlexItem>
-                    <EuiFlexItem grow={false}>
-                      <EuiButtonEmpty
-                        iconType="importAction"
-                        onClick={async () => await this.downloadCsv()}
-                      >
-                        Export formatted
-                      </EuiButtonEmpty>
-                    </EuiFlexItem>
-                    <EuiFlexItem grow={false}>
-                      <EuiButtonEmpty
-                        iconType="refresh"
-                        onClick={() => this.loadScaPolicy(this.state.lookingPolicy)}
-                      >
-                        Refresh
-                      </EuiButtonEmpty>
-                    </EuiFlexItem>
                   </EuiFlexGroup>
                   <EuiSpacer size="m" />
                   <EuiFlexGroup>
@@ -754,24 +493,9 @@ export class Inventory extends Component {
                   <EuiSpacer size="m" />
                   <EuiFlexGroup>
                     <EuiFlexItem>
-                      <TableWzAPI
-                        title="Checks"
-                        tableColumns={this.columnsChecks}
-                        tableInitialSortingField="id"
-                        searchTable={true}
-                        searchBarSuggestions={this.suggestions[this.state.lookingPolicy.policy_id]}
-                        endpoint={`/sca/${this.props.agent.id}/checks/${this.state.lookingPolicy.policy_id}`}
-                        rowProps={getChecksRowProps}
-                        tableProps={{
-                          isExpandable: true,
-                          itemIdToExpandedRowMap: this.state.itemIdToExpandedRowMap,
-                          itemId: 'id',
-                        }}
-                        downloadCsv
-                        showReload
-                        filters={this.state.filters}
-                        onFiltersChange={(filters) => this.setState({ filters })}
-                        tablePageSizeOptions={[10, 25, 50, 100]}
+                      <InventoryPolicyChecksTable
+                        agent={this.props.agent}
+                        lookingPolicy={this.state.lookingPolicy}
                       />
                     </EuiFlexItem>
                   </EuiFlexGroup>
