+++ public/components/agents/sca/inventory/index.ts
@@ -1,0 +1,2 @@
+export * from './lib';
+export * from './checks-table';
