+++ public/components/agents/sca/inventory/lib/index.ts
@@ -1,0 +1,1 @@
+export * from './api-request'
