+++ public/components/wz-search-bar/lib/base-handler.ts
@@ -12,27 +12,26 @@
 
 import { suggestItem } from '../wz-search-bar';
 
-
 export abstract class BaseHandler {
   inputStage?: string;
   isSearch?: boolean;
-  
-  async buildSuggestItems(inputValue: string):Promise<suggestItem[]> {
-    return [];
-  }
-  
-  buildSuggestFields(inputValue:string):suggestItem[] {
-    return [];
-  };
 
-  async buildSuggestValues(inputValue:string, value:string=''):Promise<suggestItem[]> {
+  async buildSuggestItems(inputValue: string): Promise<suggestItem[]> {
     return [];
   }
 
-  buildSuggestValue(value:string|number) {
+  buildSuggestFields(inputValue: string): suggestItem[] {
+    return [];
+  }
+
+  async buildSuggestValues(inputValue: string, value: string = ''): Promise<suggestItem[]> {
+    return [];
+  }
+
+  buildSuggestValue(value: string | number = '') {
     return {
-      type: {iconType: 'kqlValue', color: 'tint0'},
-      label: typeof value !== 'string' ? value.toString(): value,
+      type: { iconType: 'kqlValue', color: 'tint0' },
+      label: typeof value !== 'string' ? value.toString() : value,
     };
   }
 
@@ -40,7 +39,7 @@ export abstract class BaseHandler {
     return item.label.includes(field);
   }
 
-  filterSuggestValues(item:number|string, inputValue:string) {
+  filterSuggestValues(item: number | string, inputValue: string) {
     if (typeof item === 'number' && !!inputValue) {
       // @ts-ignore
       return item == inputValue;
@@ -51,22 +50,28 @@ export abstract class BaseHandler {
     return true;
   }
 
-  onItemClick(item:suggestItem, inputValue:string, filters:object):void {}
+  onItemClick(item: suggestItem, inputValue: string, filters: object): void {}
 
-  onInputChange(inputValue:string, currentFilters:object):{
-    isInvalid: boolean, filters: object
-  } { return {isInvalid:true, filters:{}} }
+  onInputChange(
+    inputValue: string,
+    currentFilters: object
+  ): {
+    isInvalid: boolean;
+    filters: object;
+  } {
+    return { isInvalid: true, filters: {} };
+  }
 
-  onKeyPress(inputValue:string, currentFilters:object):void {}
+  onKeyPress(inputValue: string, currentFilters: object): void {}
 
   mapSuggestFields(item): suggestItem {
     const suggestItem = {
       type: {
         iconType: 'kqlField',
-        color: 'tint4'
+        color: 'tint4',
       },
       label: item.label,
-    }
+    };
     if (item.description) {
       suggestItem['description'] = item.description;
     }
